feat: add `sourceCode.isGlobalReference(node)` method
