+++ lib/languages/js/source-code/source-code.js
@@ -8,7 +8,10 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const { isCommentToken } = require("@eslint-community/eslint-utils"),
+const {
+		isCommentToken,
+		findVariable,
+	} = require("@eslint-community/eslint-utils"),
 	TokenStore = require("./token-store"),
 	astUtils = require("../../../shared/ast-utils"),
 	Traverser = require("../../../shared/traverser"),
@@ -866,6 +869,22 @@ class SourceCode extends TokenStore {
 		}
 
 		return ancestorsStartingAtParent.reverse();
+	}
+
+	/**
+	 * Determines whether the given identifier node is a reference to a global variable.
+	 * @param {ASTNode} node `Identifier` node to check.
+	 * @returns {boolean} True if the identifier is a reference to a global variable.
+	 */
+	isGlobalReference(node) {
+		const scope = this.getScope(node);
+		const variable = findVariable(scope, node);
+
+		return (
+			variable !== null &&
+			variable.scope.type === "global" &&
+			variable.defs.length === 0
+		);
 	}
 
 	/**
