+++ lib/rules/no-promise-executor-return.js
@@ -9,7 +9,6 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const { findVariable } = require("@eslint-community/eslint-utils");
 const astUtils = require("./utils/ast-utils");
 
 //------------------------------------------------------------------------------
@@ -22,42 +21,12 @@ const functionTypesToCheck = new Set([
 ]);
 
 /**
- * Determines whether the given identifier node is a reference to a global variable.
- * @param {ASTNode} node `Identifier` node to check.
- * @param {Scope} scope Scope to which the node belongs.
- * @returns {boolean} True if the identifier is a reference to a global variable.
- */
-function isGlobalReference(node, scope) {
-	const variable = findVariable(scope, node);
-
-	return (
-		variable !== null &&
-		variable.scope.type === "global" &&
-		variable.defs.length === 0
-	);
-}
-
-/**
- * Finds function's outer scope.
- * @param {Scope} scope Function's own scope.
- * @returns {Scope} Function's outer scope.
- */
-function getOuterScope(scope) {
-	const upper = scope.upper;
-
-	if (upper.type === "function-expression-name") {
-		return upper.upper;
-	}
-	return upper;
-}
-
-/**
  * Determines whether the given function node is used as a Promise executor.
  * @param {ASTNode} node The node to check.
- * @param {Scope} scope Function's own scope.
+ * @param {SourceCode} sourceCode Source code to which the node belongs.
  * @returns {boolean} `true` if the node is a Promise executor.
  */
-function isPromiseExecutor(node, scope) {
+function isPromiseExecutor(node, sourceCode) {
 	const parent = node.parent;
 
 	return (
@@ -65,7 +34,7 @@ function isPromiseExecutor(node, scope) {
 		parent.arguments[0] === node &&
 		parent.callee.type === "Identifier" &&
 		parent.callee.name === "Promise" &&
-		isGlobalReference(parent.callee, getOuterScope(scope))
+		sourceCode.isGlobalReference(parent.callee)
 	);
 }
 
@@ -203,7 +172,7 @@ module.exports = {
 					upper: funcInfo,
 					shouldCheck:
 						functionTypesToCheck.has(node.type) &&
-						isPromiseExecutor(node, sourceCode.getScope(node)),
+						isPromiseExecutor(node, sourceCode),
 				};
 
 				if (
