+++ lib/rules/no-setter-return.js
@@ -10,33 +10,16 @@
 //------------------------------------------------------------------------------
 
 const astUtils = require("./utils/ast-utils");
-const { findVariable } = require("@eslint-community/eslint-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
 
 /**
- * Determines whether the given identifier node is a reference to a global variable.
- * @param {ASTNode} node `Identifier` node to check.
- * @param {Scope} scope Scope to which the node belongs.
- * @returns {boolean} True if the identifier is a reference to a global variable.
- */
-function isGlobalReference(node, scope) {
-	const variable = findVariable(scope, node);
-
-	return (
-		variable !== null &&
-		variable.scope.type === "global" &&
-		variable.defs.length === 0
-	);
-}
-
-/**
  * Determines whether the given node is an argument of the specified global method call, at the given `index` position.
  * E.g., for given `index === 1`, this function checks for `objectName.methodName(foo, node)`, where objectName is a global variable.
  * @param {ASTNode} node The node to check.
- * @param {Scope} scope Scope to which the node belongs.
+ * @param {SourceCode} sourceCode Source code to which the node belongs.
  * @param {string} objectName Name of the global object.
  * @param {string} methodName Name of the method.
  * @param {number} index The given position.
@@ -44,7 +27,7 @@ function isGlobalReference(node, scope) {
  */
 function isArgumentOfGlobalMethodCall(
 	node,
-	scope,
+	sourceCode,
 	objectName,
 	methodName,
 	index,
@@ -59,9 +42,8 @@ function isArgumentOfGlobalMethodCall(
 			objectName,
 			methodName,
 		) &&
-		isGlobalReference(
+		sourceCode.isGlobalReference(
 			astUtils.skipChainExpression(callNode.callee).object,
-			scope,
 		)
 	);
 }
@@ -69,21 +51,21 @@ function isArgumentOfGlobalMethodCall(
 /**
  * Determines whether the given node is used as a property descriptor.
  * @param {ASTNode} node The node to check.
- * @param {Scope} scope Scope to which the node belongs.
+ * @param {SourceCode} sourceCode Source code to which the node belongs.
  * @returns {boolean} `true` if the node is a property descriptor.
  */
-function isPropertyDescriptor(node, scope) {
+function isPropertyDescriptor(node, sourceCode) {
 	if (
 		isArgumentOfGlobalMethodCall(
 			node,
-			scope,
+			sourceCode,
 			"Object",
 			"defineProperty",
 			2,
 		) ||
 		isArgumentOfGlobalMethodCall(
 			node,
-			scope,
+			sourceCode,
 			"Reflect",
 			"defineProperty",
 			2,
@@ -101,14 +83,14 @@ function isPropertyDescriptor(node, scope) {
 			grandparent.type === "ObjectExpression" &&
 			(isArgumentOfGlobalMethodCall(
 				grandparent,
-				scope,
+				sourceCode,
 				"Object",
 				"create",
 				1,
 			) ||
 				isArgumentOfGlobalMethodCall(
 					grandparent,
-					scope,
+					sourceCode,
 					"Object",
 					"defineProperties",
 					1,
@@ -124,10 +106,10 @@ function isPropertyDescriptor(node, scope) {
 /**
  * Determines whether the given function node is used as a setter function.
  * @param {ASTNode} node The node to check.
- * @param {Scope} scope Scope to which the node belongs.
+ * @param {SourceCode} sourceCode Source code to which the node belongs.
  * @returns {boolean} `true` if the node is a setter.
  */
-function isSetter(node, scope) {
+function isSetter(node, sourceCode) {
 	const parent = node.parent;
 
 	if (
@@ -144,28 +126,13 @@ function isSetter(node, scope) {
 		parent.value === node &&
 		astUtils.getStaticPropertyName(parent) === "set" &&
 		parent.parent.type === "ObjectExpression" &&
-		isPropertyDescriptor(parent.parent, scope)
+		isPropertyDescriptor(parent.parent, sourceCode)
 	) {
 		// Setter in a property descriptor
 		return true;
 	}
 
 	return false;
-}
-
-/**
- * Finds function's outer scope.
- * @param {Scope} scope Function's own scope.
- * @returns {Scope} Function's outer scope.
- */
-function getOuterScope(scope) {
-	const upper = scope.upper;
-
-	if (upper.type === "function-expression-name") {
-		return upper.upper;
-	}
-
-	return upper;
 }
 
 //------------------------------------------------------------------------------
@@ -200,11 +167,9 @@ module.exports = {
 		 * @returns {void}
 		 */
 		function enterFunction(node) {
-			const outerScope = getOuterScope(sourceCode.getScope(node));
-
 			funcInfo = {
 				upper: funcInfo,
-				isSetter: isSetter(node, outerScope),
+				isSetter: isSetter(node, sourceCode),
 			};
 		}
