+++ lib/rules/prefer-regex-literals.js
@@ -14,7 +14,6 @@ const {
 	CALL,
 	CONSTRUCT,
 	ReferenceTracker,
-	findVariable,
 } = require("@eslint-community/eslint-utils");
 const {
 	RegExpValidator,
@@ -168,22 +167,6 @@ module.exports = {
 		const sourceCode = context.sourceCode;
 
 		/**
-		 * Determines whether the given identifier node is a reference to a global variable.
-		 * @param {ASTNode} node `Identifier` node to check.
-		 * @returns {boolean} True if the identifier is a reference to a global variable.
-		 */
-		function isGlobalReference(node) {
-			const scope = sourceCode.getScope(node);
-			const variable = findVariable(scope, node);
-
-			return (
-				variable !== null &&
-				variable.scope.type === "global" &&
-				variable.defs.length === 0
-			);
-		}
-
-		/**
 		 * Determines whether the given node is a String.raw`` tagged template expression
 		 * with a static template literal.
 		 * @param {ASTNode} node Node to check.
@@ -193,7 +176,7 @@ module.exports = {
 			return (
 				node.type === "TaggedTemplateExpression" &&
 				astUtils.isSpecificMemberAccess(node.tag, "String", "raw") &&
-				isGlobalReference(
+				sourceCode.isGlobalReference(
 					astUtils.skipChainExpression(node.tag).object,
 				) &&
 				astUtils.isStaticTemplateLiteral(node.quasi)
